simurgh/requirements.txt
pymssql
psycopg2-binary
pandas
sqlalchemy